dash/requirements.txt
# ── Core API ──────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pyyaml>=6.0.1
httpx>=0.27.0

# ── Messaging ─────────────────────────────────────────────────────────────────
aiokafka>=0.10.0          # Kafka (optional — in-memory bus used when Kafka unavailable)

# ── Collectors ────────────────────────────────────────────────────────────────
feedparser>=6.0.11         # RSS / Atom feed parsing
yt-dlp>=2024.4.9           # YouTube download and search
youtube-transcript-api>=0.6.2  # YouTube captions (fast, no download)
# snscrape                 # Twitter scraping — install manually: pip install snscrape
# instaloader              # Instagram scraping — install manually: pip install instaloader

# ── Processors ───────────────────────────────────────────────────────────────
spacy>=3.7.0               # NER — also run: python -m spacy download en_core_web_sm
vaderSentiment>=3.3.2      # Sentiment (social-media optimized)
textblob>=0.18.0           # Sentiment (long-form text)
sentence-transformers>=2.7.0  # Embeddings (all-MiniLM-L6-v2)

# Optional — uncomment when needed:
# faster-whisper>=1.0.0    # Local Whisper STT
# pytesseract>=0.3.10      # OCR wrapper (requires tesseract system install)

# ── Consumers ────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9     # PostgreSQL
qdrant-client>=1.9.0       # Qdrant vector DB

# ── Dev / testing ─────────────────────────────────────────────────────────────
requests>=2.31.0
